Split app initiation from config into a separate file
